Adding a signal that will be sent out once the petition has been saved in the create_new_petition view.
This signal will carry with it the bound form as well as the newly created petition, for any post processing (
such as operating on additional, non model related fields, defined in the ModelForm).

This allows me to make the CreatePetitionForm a vanilla ModelForm. Any additional customizations can be done using a custom form passed into the view.